Add skeletons of base unit and measurement classes.

+++ src/test/java/com/ford/openxc/UnitsTest.java
@@ -1,11 +1,13 @@
 package com.ford.openxc;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.instanceOf;
+import static org.hamcrest.Matchers.equalTo;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+
+import com.ford.openxc.units.Meter;
 
 public class UnitsTest {
 
@@ -20,6 +22,6 @@ public class UnitsTest {
     @Test
     public void testComparableToDouble() {
         Meter value = new Meter(10);
-        assertThat(value, equalTo(10));
+        assertThat(value.doubleValue(), equalTo(10.0));
     }
 }
